requirements.txt
python-telegram-bot==22.6
python-dotenv==1.0.1